+++ app/Policies/WorkoutPolicy.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace App\Policies;
 
-use App\Models\Repetition;
 use App\Models\User;
+use App\Models\Workout;
 use Illuminate\Auth\Access\HandlesAuthorization;
 
-class RepetitionPolicy
+class WorkoutPolicy
 {
     use HandlesAuthorization;
 
@@ -27,35 +27,35 @@ class RepetitionPolicy
      * Determine whether the user can update the model.
      *
      * @param  \App\Models\User  $user
-     * @param  \App\Models\Repetition  $repetition
+     * @param  \App\Models\Workout  $workout
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function update(User $user, Repetition $repetition)
+    public function update(User $user, Workout $workout)
     {
-        return $repetition->user_id == $user->id;
+        return ! $workout->trashed() && ($user->is_admin || $workout->user_id == $user->id);
     }
 
     /**
      * Determine whether the user can delete the model.
      *
      * @param  \App\Models\User  $user
-     * @param  \App\Models\Repetition  $repetition
+     * @param  \App\Models\Workout  $workout
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function delete(User $user, Repetition $repetition)
+    public function delete(User $user, Workout $workout)
     {
-        return $repetition->user_id == $user->id;
+        return ! $workout->trashed() && ($user->is_admin || $workout->user_id == $user->id);
     }
 
     /**
      * Determine whether the user can restore the model.
      *
      * @param  \App\Models\User  $user
-     * @param  \App\Models\Repetition  $repetition
+     * @param  \App\Models\Workout  $workout
      * @return \Illuminate\Auth\Access\Response|bool
      */
-    public function restore(User $user, Repetition $repetition)
+    public function restore(User $user, Workout $workout)
     {
-        return false;
+        return $workout->trashed() && $user->is_admin;
     }
 }