+++ resources/lang/en/resources.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+
+    'exercise' => [
+        'store' => 'Exercise details have been successfully stored!',
+        'update' => 'Exercise details have been successfully changed!',
+        'destroy' => 'Exercise has been successfully deleted!',
+    ],
+
+];