+++ resources/views/components/form/select.blade.php
@@ -1,0 +1,10 @@
+<div class="form-floating mb-3">
+    <select id="{{ $name }}" name="{{ $name }}"
+            class="form-control text-light bg-dark @error($name) is-invalid @enderror">
+        {{ $slot }}
+    </select>
+    <label for="{{ $name }}">{{ $label }}</label>
+    @error($name)
+    <div class="d-block invalid-feedback">{{ $message }}</div>
+    @enderror
+</div>