+++ CHANGELOG.md
@@ -9,12 +9,12 @@ All notable changes to this project will be documented in this file.
 
 * Change password ([fbad79](https://github.com/kishieel/gym-tracker-website/commit/fbad791db470bdac38b9eb207c38f3aae05995f1))
 * Convetional changelog ([c3b348](https://github.com/kishieel/gym-tracker-website/commit/c3b3487dda18fcd0d65b978ed2ddaf2eb583f65e))
-* Exercise repetition create/edit form, policies, leaderboard, history ([5d398a](https://github.com/kishieel/gym-tracker/commit/5d398a8aa709a2bba5eb75075c6e6fd001bfcc3c))
+* Exercise workout create/edit form, policies, leaderboard, history ([5d398a](https://github.com/kishieel/gym-tracker/commit/5d398a8aa709a2bba5eb75075c6e6fd001bfcc3c))
 
 ### Bug Fixes
 
 * Production deployment ([40eb24](https://github.com/kishieel/gym-tracker-website/commit/40eb24fe84933867c3efc5dfb92f0fb8aad34668))
-* Repetition policy ([de03c4](https://github.com/kishieel/gym-tracker-website/commit/de03c4a6a85a8f879149b76b62d8f353e06a7bea))
+* Workout policy ([de03c4](https://github.com/kishieel/gym-tracker-website/commit/de03c4a6a85a8f879149b76b62d8f353e06a7bea))
 
 ### Continuous Integrations